[Metricbeat] Migrate HTTP json Metricset to use ReporterV2 interface (#10960)
