+++ metricbeat/module/http/json/_meta/data.json
@@ -1,18 +1,20 @@
 {
-    "@timestamp": "2017-10-12T08:05:34.853Z",
-    "beat": {
-        "hostname": "host.example.com",
-        "name": "host.example.com"
+    "@timestamp": "2019-03-01T08:05:34.853Z",
+    "event": {
+        "dataset": "http.test",
+        "duration": 115000,
+        "module": "http"
     },
     "http": {
-        "json": {
+        "test": {
             "hello": "world"
         }
     },
     "metricset": {
-        "host": "127.0.0.1:8080",
-        "module": "http",
-        "name": "json",
-        "rtt": 115
+        "name": "json"
+    },
+    "service": {
+        "address": "127.0.0.1:55555",
+        "type": "http"
     }
 }
