+++ metricbeat/module/http/json/_meta/testdata/config.yml
@@ -1,0 +1,6 @@
+type: http
+url: "/"
+module:
+  namespace: test
+omit_documented_fields_check:
+  - "http.test.*"
