+++ metricbeat/module/http/json/_meta/testdata/docs.json
@@ -1,0 +1,3 @@
+{
+    "hello": "world"
+}
