+++ metricbeat/module/http/json/_meta/testdata/docs.json-expected.json
@@ -1,0 +1,21 @@
+[
+    {
+        "event": {
+            "dataset": "http.test",
+            "duration": 115000,
+            "module": "http"
+        },
+        "http": {
+            "test": {
+                "hello": "world"
+            }
+        },
+        "metricset": {
+            "name": "json"
+        },
+        "service": {
+            "address": "127.0.0.1:55555",
+            "type": "http"
+        }
+    }
+]
