+++ metricbeat/module/http/json/json.go
@@ -20,9 +20,8 @@ package json
 import (
 	"encoding/json"
 	"io/ioutil"
-	"net/http"
-	"strconv"
-	"strings"
+
+	"github.com/pkg/errors"
 
 	"github.com/elastic/beats/libbeat/common"
 	"github.com/elastic/beats/metricbeat/helper"
@@ -116,95 +115,52 @@ func New(base mb.BaseMetricSet) (mb.MetricSet, error) {
 	}, nil
 }
 
-func (m *MetricSet) processBody(response *http.Response, jsonBody interface{}) common.MapStr {
-	var event common.MapStr
-
-	if m.deDotEnabled {
-		event = common.DeDotJSON(jsonBody).(common.MapStr)
-	} else {
-		event = jsonBody.(common.MapStr)
-	}
-
-	if m.requestEnabled {
-		event[mb.ModuleDataKey] = common.MapStr{
-			"request": common.MapStr{
-				"headers": m.getHeaders(response.Request.Header),
-				"method":  response.Request.Method,
-				"body": common.MapStr{
-					"content": m.body,
-				},
-			},
-		}
-	}
-
-	if m.responseEnabled {
-		phrase := strings.TrimPrefix(response.Status, strconv.Itoa(response.StatusCode)+" ")
-		event[mb.ModuleDataKey] = common.MapStr{
-			"response": common.MapStr{
-				"code":    response.StatusCode,
-				"phrase":  phrase,
-				"headers": m.getHeaders(response.Header),
-			},
-		}
-	}
-
-	// Set dynamic namespace
-	event["_namespace"] = m.namespace
-
-	return event
-}
-
-// Fetch methods implements the data gathering and data conversion to the right format
-// It returns the event which is then forward to the output. In case of an error, a
-// descriptive error must be returned.
-func (m *MetricSet) Fetch() ([]common.MapStr, error) {
+// Fetch methods implements the data gathering and data conversion to the right
+// format. It publishes the event which is then forwarded to the output. In case
+// of an error set the Error field of mb.Event or simply call report.Error().
+func (m *MetricSet) Fetch(reporter mb.ReporterV2) error {
 	response, err := m.http.FetchResponse()
 	if err != nil {
-		return nil, err
+		return err
 	}
-	defer response.Body.Close()
-
-	var jsonBody common.MapStr
-	var jsonBodyArr []common.MapStr
-	var events []common.MapStr
+	defer func() {
+		if err := response.Body.Close(); err != nil {
+			m.Logger().Debug("error closing http body")
+		}
+	}()
 
 	body, err := ioutil.ReadAll(response.Body)
 	if err != nil {
-		return nil, err
+		return err
 	}
 
 	if m.jsonIsArray {
-		err = json.Unmarshal(body, &jsonBodyArr)
-		if err != nil {
-			return nil, err
+		var jsonBodyArr []common.MapStr
+		if err = json.Unmarshal(body, &jsonBodyArr); err != nil {
+			return err
 		}
 
 		for _, obj := range jsonBodyArr {
 			event := m.processBody(response, obj)
-			events = append(events, event)
+
+			if reported := reporter.Event(event); !reported {
+				m.Logger().Debug(errors.Errorf("error reporting event: %#v", event))
+				return nil
+			}
 		}
 	} else {
-		err = json.Unmarshal(body, &jsonBody)
-		if err != nil {
-			return nil, err
+		var jsonBody common.MapStr
+		if err = json.Unmarshal(body, &jsonBody); err != nil {
+			return err
 		}
 
 		event := m.processBody(response, jsonBody)
-		events = append(events, event)
-	}
 
-	return events, nil
-}
-
-func (m *MetricSet) getHeaders(header http.Header) map[string]string {
-	headers := make(map[string]string)
-	for k, v := range header {
-		value := ""
-		for _, h := range v {
-			value += h + " ,"
+		if reported := reporter.Event(event); !reported {
+			m.Logger().Debug(errors.Errorf("error reporting event: %#v", event))
+			return nil
 		}
-		value = strings.TrimRight(value, " ,")
-		headers[k] = value
 	}
-	return headers
+
+	return nil
 }
